+++ src/GitVersionCore/Core/Abstractions/Git/IBranch.cs
