+++ src/GitVersionCore/Core/Abstractions/Git/ICommit.cs
