+++ src/GitVersionCore/Core/Abstractions/Git/IObjectId.cs
