+++ src/GitVersionCore/Core/Abstractions/Git/IReference.cs
