+++ src/GitVersionCore/Core/Abstractions/Git/IRemote.cs
