+++ src/GitVersionCore/Core/Abstractions/Git/ITag.cs
